refactor structure with focus on memory and dynamics
