+++ crates/evorl-core/src/environment.rs
@@ -1,8 +1,4 @@
-use crate::action::Action;
-use crate::dynamics::{ExperienceTuple, History, HistoryRepresentation, Reward, UpdateFunction};
-use crate::state::{Observation, State};
-//use crate::tensor_convert::TensorConvertible;
-//use std::collections::VecDeque;
+use crate::base::{Action, Observation, Reward, State};
 use std::fmt::Debug;
 
 /// Error type for environment operations.
