+++ crates/evorl-core/src/lib.rs
@@ -21,10 +21,10 @@
 
 pub mod action;
 pub mod agent;
-pub mod dynamics;
+pub mod base;
 pub mod environment;
 pub mod evolution;
+pub mod experience;
 pub mod memory;
 pub mod metrics;
 pub mod state;
-pub mod tensor_convert;
