+++ crates/evorl-envs/src/games/chess/moves.rs
@@ -104,8 +104,8 @@
 //!   environment. Agents learn to avoid them through reward signals.
 
 use crate::games::chess::board::Square;
-use evorl_core::action::{Action, MultiDiscreteAction};
-use evorl_core::tensor_convert::TensorConvertible;
+use evorl_core::action::MultiDiscreteAction;
+use evorl_core::base::{Action, TensorConvertible};
 
 /// Promotion piece types for pawn promotion moves.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
